requirements.txt
openai
langchain
fastapi
uvicorn
python-dotenv
